+++ ouds_core/lib/components/utilities/badge_border_utils.dart
@@ -10,6 +10,7 @@
  * // Software description: Flutter library of reusable graphical components
  * //
  */
+
 /// @nodoc
 library;